+++ testapps/GHIssues/src/main/java/ghissues/Gh797.java
@@ -2,14 +2,13 @@ package ghissues;
 
 import act.controller.annotation.UrlContext;
 import act.util.JsonView;
-import act.util.LogSupport;
 import org.osgl.mvc.annotation.GetAction;
 
 import java.util.Date;
 
 @UrlContext("797")
 @JsonView
-public class Gh797 extends LogSupport {
+public class Gh797 extends BaseController {
 
     public static class Foo {
         public Date date = new Date();