+++ testapps/GHIssues/src/main/java/ghissues/Gh798.java
@@ -2,7 +2,6 @@ package ghissues;
 
 import act.controller.annotation.UrlContext;
 import act.util.JsonView;
-import act.util.LogSupport;
 import com.alibaba.fastjson.annotation.JSONField;
 import org.osgl.mvc.annotation.GetAction;
 
@@ -10,7 +9,7 @@ import java.util.Date;
 
 @UrlContext("798")
 @JsonView
-public class Gh798 extends LogSupport {
+public class Gh798 extends BaseController {
 
     public static class Foo {
         @JSONField(format = "yyyy-MM")