+++ testapps/GHIssues/src/main/java/ghissues/Gh809.java
@@ -4,11 +4,10 @@ import static act.controller.Controller.Util.template;
 
 import act.Act;
 import act.controller.annotation.UrlContext;
-import act.util.LogSupport;
 import org.osgl.mvc.annotation.GetAction;
 
 @UrlContext("809")
-public class Gh809 extends LogSupport {
+public class Gh809 extends BaseController {
 
     @GetAction
     public void index() {