+++ testapps/GHIssues/src/main/java/ghissues/Gh829.java
@@ -5,7 +5,6 @@ import act.app.ActionContext;
 import act.controller.annotation.UrlContext;
 import act.inject.param.NoBind;
 import act.job.OnAppStart;
-import act.util.LogSupport;
 import org.osgl.mvc.annotation.GetAction;
 import org.osgl.util.S;
 
@@ -13,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 @UrlContext("829")
-public class Gh829 extends LogSupport {
+public class Gh829 extends BaseController {
 
     public static class Foo {
         public String name = S.random();