+++ testapps/GHIssues/src/main/java/ghissues/Gh832.java
@@ -4,7 +4,7 @@ import act.controller.annotation.UrlContext;
 import org.osgl.mvc.annotation.GetAction;
 
 @UrlContext("832")
-public class Gh832 {
+public class Gh832 extends BaseController {
 
     @GetAction("/post/{cityId}")
     public int action(int cityId) {