+++ testapps/GHIssues/src/main/java/ghissues/Gh834.java
@@ -1,6 +1,7 @@
 package ghissues;
 
-import act.controller.Controller;
+import static act.controller.Controller.Util.renderImage;
+
 import act.controller.annotation.UrlContext;
 import org.osgl.mvc.annotation.GetAction;
 import org.osgl.util.Img;
@@ -8,7 +9,7 @@ import org.osgl.util.Img;
 import java.awt.image.BufferedImage;
 
 @UrlContext("834")
-public class Gh834 extends Controller.Util {
+public class Gh834 extends BaseController {
 
     @GetAction
     public void test() {