+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * `CacheFor` - make it allow skip `Cache-Control` header #837
 * `ResponseCache` - content disposition shall be cached #838
 * `CacheFor` key shall include `Accept` type #836
+* Provide a mechanism to allow developer advice on return value of request handler method #835
 * Add render image support #834
 * Optimise `LogSupport` for `DestroyableBase` #833
 * Allow app developer to specify download file name #829